requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chess-cheat==1.0.3
constantly==15.1.0
cryptography==38.0.1
cssselect==1.1.0
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
filelock==3.8.0
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
gunicorn==20.1.0
h5py==3.7.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mss==6.1.0
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
Protego==0.2.1
protobuf==3.19.5
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyscreenshot==3.0
pytz==2022.4
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
Scrapy==2.7.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stockfish==3.28.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
timeout-decorator==0.5.0
tldextract==3.4.0
Twisted==22.8.0
typing_extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.0
